+++ packages/rx_bloc_cli/bin/compile_bundles.sh
@@ -74,4 +74,10 @@ dart run mason_cli:mason bundle \
 dart run mason_cli:mason bundle \
   -t dart \
   -o lib/src/templates/ \
+  mason_templates/bricks/feature_cicd_fastlane
+
+dart run mason_cli:mason bundle \
+  -t dart \
+  -o lib/src/templates/ \
   mason_templates/bricks/lib_auth_matrix
+