+++ packages/rx_bloc_cli/lib/src/commands/create_command.dart
@@ -97,14 +97,24 @@ class CreateCommand extends Command<int> {
         'enable_patrol': arguments.patrolTestsEnabled,
         'has_authentication': arguments.authenticationEnabled,
         'realtime_communication': arguments.realtimeCommunicationEnabled,
+        'cicd': arguments.cicdEnabled,
         'enable_auth_matrix': arguments.authMatrixEnabled,
       },
     );
 
-    // Manually create gitignore.
+    // Manually create project gitignore
     GitIgnoreCreator.generate(arguments.outputDirectory.path);
 
-    final fileCount = generatedFiles.length + 1;
+    var fileCount = generatedFiles.length + 1;
+
+    // Manually create devops gitignore
+    if (arguments.cicdEnabled) {
+      GitIgnoreCreator.generate(
+        '${arguments.outputDirectory.path}/devops',
+        gitignore: GitignoreType.devops,
+      );
+      fileCount++;
+    }
 
     fileGenerationProgress.complete('Bootstrapping done');
 
@@ -206,6 +216,7 @@ class CreateCommand extends Command<int> {
     _usingLog('Patrol integration tests', arguments.patrolTestsEnabled);
     _usingLog('Realtime communication', arguments.realtimeCommunicationEnabled);
     _usingLog('Auth matrix', arguments.authMatrixEnabled);
+    _usingLog('CI/CD', arguments.cicdEnabled);
   }
 
   /// Shows a delayed log with a success symbol in front of it