+++ packages/rx_bloc_cli/lib/src/models/configurations/feature_configuration.dart
@@ -11,6 +11,7 @@ class FeatureConfiguration {
     required this.deepLinkEnabled,
     required this.devMenuEnabled,
     required this.patrolTestsEnabled,
+    required this.cicdEnabled,
     required this.authMatrixEnabled,
   });
 
@@ -40,6 +41,9 @@ class FeatureConfiguration {
 
   /// Patrol tests
   final bool patrolTestsEnabled;
+
+  /// CICD
+  final bool cicdEnabled;
 
   /// Uses Firebase
   bool get usesFirebase => analyticsEnabled || pushNotificationsEnabled;