+++ packages/rx_bloc_cli/lib/src/models/generator_arguments_provider.dart
@@ -1,6 +1,7 @@
 import 'dart:io';
 
 import 'package:mason/mason.dart';
+import 'package:rx_bloc_cli/src/models/ci_cd_type.dart';
 import 'package:rx_bloc_cli/src/models/configurations/auth_configuration.dart';
 import 'package:rx_bloc_cli/src/models/configurations/feature_configuration.dart';
 import 'package:rx_bloc_cli/src/models/generator_arguments.dart';
@@ -124,6 +125,10 @@ class GeneratorArgumentsProvider {
     // Patrol tests
     final patrolTestsEnabled = _reader.read<bool>(CommandArguments.patrol);
 
+    // CI/CD
+    final cicdType = _reader.read<CICDType>(CommandArguments.cicd);
+    final cicdEnabled = cicdType != CICDType.none;
+
     // Auth matrix
     final authMatrixEnabled = _reader.read<bool>(CommandArguments.authMatrix);
 
@@ -137,6 +142,7 @@ class GeneratorArgumentsProvider {
       deepLinkEnabled: deepLinkEnabled,
       devMenuEnabled: devMenuEnabled,
       patrolTestsEnabled: patrolTestsEnabled,
+      cicdEnabled: cicdEnabled,
       authMatrixEnabled: authMatrixEnabled,
     );
   }