+++ packages/rx_bloc_cli/test/models/generator_arguments_test.dart
@@ -36,6 +36,7 @@ void main() {
       deepLinkEnabled: true,
       devMenuEnabled: true,
       patrolTestsEnabled: true,
+      cicdEnabled: true,
       authMatrixEnabled: true,
     );
     sut = GeneratorArguments(
@@ -80,6 +81,7 @@ void main() {
           equals(featureConfiguration.realtimeCommunicationEnabled));
       expect(sut.widgetToolkitEnabled,
           equals(featureConfiguration.widgetToolkitEnabled));
+      expect(sut.cicdEnabled, equals(featureConfiguration.cicdEnabled));
     });
   });
 }